refactoring for new hyper and serde
