+++ Cargo.toml
@@ -8,24 +8,16 @@ documentation = "https://kstrempel.github.io/rust-kairosdb/kairosdb/index.html"
 description = "Client for the KairosDB REST API"
 readme = "README.md"
 keywords = ["kairosdb", "timeseries", "database"]
-build = "build.rs"
 categories = ["database"]
 
 [lib]
 name = "kairosdb"
 
-[features]
-default = ["serde_codegen"]
-unstable = ["serde_derive"]
-
-[build-dependencies]
-serde_codegen = { version = "0.8", optional = true }
-
 [dependencies]
-chrono="0.2"
-hyper="^0.9"
-log = "^0.3"
-env_logger = "0.3.5"
-serde = "0.8"
-serde_derive = { version = "0.8", optional = true }
-serde_json = "0.8"
+chrono="0.4.6"
+hyper="0.12.19"
+log = "0.4.6"
+env_logger = "0.6.0"
+serde = "1.0.84"
+serde_derive = "1.0"
+serde_json = "1.0.34"
