--- build.rs
@@ -1,40 +1,0 @@
-#[cfg(feature = "serde_codegen")]
-fn main() {
-    extern crate serde_codegen;
-
-    use std::env;
-    use std::path::Path;
-
-    let out_dir = env::var_os("OUT_DIR").unwrap();
-
-    let src = Path::new("src/datapoints.in.rs");
-    let dst = Path::new(&out_dir).join("datapoints.rs");
-
-    serde_codegen::expand(&src, &dst).unwrap();
-
-    let src = Path::new("src/lib.in.rs");
-    let dst = Path::new(&out_dir).join("lib.rs");
-
-    serde_codegen::expand(&src, &dst).unwrap();
-
-    let src = Path::new("src/query.in.rs");
-    let dst = Path::new(&out_dir).join("query.rs");
-
-    serde_codegen::expand(&src, &dst).unwrap();
-
-    let src = Path::new("src/result.in.rs");
-    let dst = Path::new(&out_dir).join("result.rs");
-
-    serde_codegen::expand(&src, &dst).unwrap();
-
-    let src = Path::new("src/helper.in.rs");
-    let dst = Path::new(&out_dir).join("helper.rs");
-
-    serde_codegen::expand(&src, &dst).unwrap();
-
-}
-
-#[cfg(not(feature = "serde_codegen"))]
-fn main() {
-    // do nothing
-}
