+++ src/datapoints.rs
@@ -12,11 +12,42 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
+use std::collections::HashMap;
+use chrono::{DateTime, TimeZone};
 
-//! Creates new datapoints
+/// Struct to define everything for a datapoint
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Datapoints {
+    name: String,
+    datapoints: Vec<(i64, f64)>,
+    tags: HashMap<String, String>,
+    ttl: u32,
+}
 
-#[cfg(feature = "serde_macros")]
-include!("datapoints.in.rs");
+impl Datapoints {
+    /// Creates a new set of datapoints
+    pub fn new(name: &str, ttl: u32) -> Datapoints {
+        Datapoints {
+            name: name.to_string(),
+            datapoints: Vec::new(),
+            tags: HashMap::new(),
+            ttl: ttl,
+        }
+    }
 
-#[cfg(not(feature = "serde_macros"))]
-include!(concat!(env!("OUT_DIR"), "/datapoints.rs"));
+    /// Adds a new datapoint to the set using 'DateTime'
+    pub fn add<Tz: TimeZone>(&mut self, datetime: DateTime<Tz>, value: f64) {
+        self.datapoints.push((datetime.timestamp() * 1000, value));
+    }
+
+    /// Adds a new datapoint to the set using the unix millisecond as
+    /// time reference
+    pub fn add_ms(&mut self, ms: i64, value: f64) {
+        self.datapoints.push((ms, value));
+    }
+
+    /// Adds a tag to the datapoint set
+    pub fn add_tag(&mut self, name: &str, value: &str) {
+        self.tags.insert(name.to_string(), value.to_string());
+    }
+}
